App hotfix 1.22.4

+++ src/nbgl/ui_utils.h
@@ -11,7 +11,7 @@ extern char *g_finishMsg;
 
 void ui_all_cleanup(void);
 
-bool ui_pairs_init(uint8_t nbPairs);
+bool ui_pairs_init(size_t nbPairs);
 void ui_pairs_cleanup(void);
 
 bool ui_buffers_init(uint8_t title_len, uint8_t subtitle_len, uint8_t finish_len);
